refactor: remove  from the  interface
